updates to update

+++ js/removing.js
@@ -11,33 +11,38 @@ class User {
         return new Promise((accept, deny) => {
             db.all("select * from `token` where `name` = $name and `token` = $token;", { $name: this.pname, $token: this.ptoken }, (err, res) => {
                 if (res[0]) {
-                    this.remove_user();
-                    accept({ status: true, name: this.pname, token: this.ptoken });
+                    accept(this.remove_user(true));
                 }
                 else {
-                    accept({ status: false, name: this.pname, token: "Token not Found" });
+                    accept(this.remove_user(false));
                 }
             });
         });
     }
-    remove_user() {
-        db.run("DELETE FROM `users` WHERE `name` = $name;", { $name: this.pname }, (err, res) => {
-            if (err) {
-                console.log(err);
-            }
-            else {
-                console.log("Sucess on Deletion");
-            }
-        });
-        db.run("DELETE FROM `token` WHERE $name = name and `token` = $token;", { $name: this.pname, $token: this.ptoken }, (err, res) => {
-            if (err) {
-                console.log(err);
-            }
-            else {
-                console.log("Sucess on token deletion");
-            }
-            ;
-        });
+    remove_user(token_found) {
+        if (token_found) {
+            db.run("DELETE FROM `users` WHERE `name` = $name;", { $name: this.pname }, (err, res) => {
+                if (err) {
+                    console.log(err);
+                }
+                else {
+                    console.log("Sucess on Deletion");
+                }
+            });
+            db.run("DELETE FROM `token` WHERE $name = name and `token` = $token;", { $name: this.pname, $token: this.ptoken }, (err, res) => {
+                if (err) {
+                    console.log(err);
+                }
+                else {
+                    console.log("Sucess on token deletion");
+                }
+                ;
+            });
+            return { status: true, secret_token: this.ptoken, name: this.pname };
+        }
+        else {
+            return { status: false };
+        }
     }
 }
 exports.default = User;
